Updated UI_backpack in MainFloor(7)
fixed previous bug. Now UI updates when: Esc and checked UI, then Esc to go back to the game and add more item/info, then Esc and check the UI again(in the previous version it won't update this time).

now only the mainfloor has this backpack item/info UI. The rest of the scenes are still using dummy text.